Bayesian Polya-gamma regression

this function runs the Bayesian multinomial regression using Polya-gamma data augmentation

pg_splm_mra(
  Y,
  X,
  locs,
  params,
  priors,
  n_cores = 1L,
  M = 4,
  n_coarse_grid = 10,
  inits = NULL,
  config = NULL,
  n_chain = 1,
  progress = FALSE,
  verbose = FALSE,
  use_spam = TRUE
)

Arguments

Y

is a \(n \times J\) matrix of compositional count data.

X

is a \(n \times p\) matrix of climate variables.

locs

is a \(n \times 2\) matrix of observation locations.

params

is a list of parameter settings. The list params must contain the following values:

n_adapt: A positive integer number of adaptive MCMC iterations.

n_mcmc: A positive integer number of total MCMC iterations post adaptation.

n_thin: A positive integer number of MCMC iterations per saved sample.

n_message: A positive integer number of frequency of iterations to output a progress message. For example, n_message = 50 outputs progress messages every 50 iterations.

priors

is the list of prior settings.

n_cores

is the number of cores for parallel computation using openMP.

M

The number of resolutions.

n_coarse_grid

The number of basis functions in one direction (e.g. n_coarse_grid = 10 results in a \(10 \times 10\) course grid which is further extended by the number of additional padding basis functions given by n_padding.

inits

is the list of initial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.

config

is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.

n_chain

is the MCMC chain id. The default is 1.

progress

is a logical input that determines whether to print a progress bar.

verbose

is a logical input that determines whether to print more detailed messages.

use_spam

is a boolean flag to determine whether the output is a list of spam matrix objects (use_spam = TRUE) or a an \(n \times n\) sparse Matrix of class "dgCMatrix" use_spam = FALSE (see spam and Matrix packages for details).