Bayesian Polya-gamma regression

this function runs the Bayesian multinomial regression using Polya-gamma data augmentation

pg_stlm(
  Y,
  X,
  locs,
  params,
  priors,
  corr_fun = "exponential",
  n_cores = 1L,
  shared_covariance_params = TRUE,
  inits = NULL,
  config = NULL,
  n_chain = 1,
  progress = FALSE,
  verbose = FALSE
)

Arguments

Y

is a \(N \times J \times T\) array of compositional count data.

X

is a \(N \times p\) matrix of climate variables.

locs

is a \(n_sites \times 2\) matrix of observation locations.

params

is a list of parameter settings. The list params must contain the following values:

n_adapt: A positive integer number of adaptive MCMC iterations.

n_mcmc: A positive integer number of total MCMC iterations post adaptation.

n_thin: A positive integer number of MCMC iterations per saved sample.

n_message: A positive integer number of frequency of iterations to output a progress message. For example, n_message = 50 outputs progress messages every 50 iterations.

priors

is a list of prior settings.

corr_fun

is a character that denotes the correlation function form. Current options include "matern" and "exponential".

n_cores

is the number of cores for parallel computation using openMP.

shared_covariance_params

is a logicial input that determines whether to fit the spatial process with component specifice parameters. If TRUE, each component has conditionally independent Gaussian process parameters theta and tau2. If FALSE, all components share the same Gaussian process parameters theta and tau2.

inits

is the list of intial values if the user wishes to specify initial values. If these values are not specified, then the intital values will be randomly sampled from the prior.

config

is the list of configuration values if the user wishes to specify initial values. If these values are not specified, then default a configuration will be used.

n_chain

is the MCMC chain id. The default is 1.

progress

is a logicial input that determines whether to print a progress bar.

verbose

is a logicial input that determines whether to print more detailed messages.