Bayesian Polya-gamma regression prediction

this function generates predictions from the Bayesian multinomial regression using Polya-gamma data augmentation

predict_pg_splm_mra(
  out,
  X,
  X_pred,
  locs,
  locs_pred,
  corr_fun,
  shared_covariance_params,
  progress = TRUE,
  verbose = FALSE,
  force = FALSE
)

Arguments

out

is a list of MCMC outputs from pg_splm()

X

is a \(n \times p\) matrix of covariates at the observed locations.

X_pred

is a \(n_{pred} \times p\) matrix of covariates at the locations where predictions are to be made.

locs

is a \(n \times 2\) matrix of locations where observations were taken.

locs_pred

is a \(n_pred \times 2\) matrix of locations where predictions are to be made.

corr_fun

is a character that denotes the correlation function form. Current options include "matern" and "exponential".

shared_covariance_params

is a logicial input that determines whether to fit the spatial process with component specifice parameters. If TRUE, each component has conditionally independent Gaussian process parameters theta and tau2. If FALSE, all components share the same Gaussian process parameters theta and tau2.

progress

is a logicial input that determines whether to print a progress bar.

verbose

is a logicial input that determines whether to print more detailed messages.

force

is a logicial input that determines whether to allow for predictions at more locations than 10000. The default is FALSE